+++ src/index.js
@@ -1,185 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css'
+import Game from './components/Game';
+import './index.css';
 
 
-function Square(props) {
-    return (
-        <button className="square" onClick={props.onClick}>
-            {props.value}
-        </button>
-    )
-}
-
-function calculateWinner(squares) {
-    const winnerLines = [
-        [0, 1, 2],
-        [3, 4, 5],
-        [6, 7, 8],
-        [0, 3, 6],
-        [1, 4, 7],
-        [2, 5, 8],
-        [0, 4, 8],
-        [2, 4, 6],
-    ];
-
-    for (let i = 0; i < winnerLines.length; i++) {
-        const [a, b, c] = winnerLines[i];
-        if (squares[a] && squares[a] === squares[b] && squares[a] === squares[c])
-            return squares[a];
-    }
-    return null;
-}
 
 
-class Board extends React.Component {
-    renderSquare(i) {
-        return (
-            <Square value={this.props.squares[i]}
-                onClick={() => this.props.onClick(i)}
-            />
-        );
-    }
 
-    render() {
-        return (
-            <div>
-                <div className="board-row">
-                    {this.renderSquare(0)}
-                    {this.renderSquare(1)}
-                    {this.renderSquare(2)}
-                </div>
-                <div className="board-row">
-                    {this.renderSquare(3)}
-                    {this.renderSquare(4)}
-                    {this.renderSquare(5)}
-                </div>
-                <div className="board-row">
-                    {this.renderSquare(6)}
-                    {this.renderSquare(7)}
-                    {this.renderSquare(8)}
-                </div>
-            </div>
-        )
-    }
-}
-
-
-class NameForm extends React.Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            playerOne: '',
-            playerTwo: ''
-        };
-        this.handlePlayerOneChange = this.handlePlayerOneChange.bind(this);
-        this.handlePlayerTwoChange = this.handlePlayerTwoChange.bind(this);
-        this.handleSubmit = this.handleSubmit.bind(this);
-    }
-
-    handlePlayerOneChange(event) { this.setState({ playerOne: event.target.value }); }
-    handlePlayerTwoChange(event) { this.setState({ playerTwo: event.target.value }); }
-    handleSubmit(event) {
-        alert('A name was submitted: \n player 1 : ' + this.state.playerOne + " \n player 2: " + this.state.playerTwo);
-        event.preventDefault();
-    }
-
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-                <label>
-                    Player 1:
-                    <input type="text" value={this.state.playerOne} onChange={this.handlePlayerOneChange} />
-                </label>
-                <br/><br/>
-                <label>
-                    Player 2:
-                    <input type="text" value={this.state.playerTwo} onChange={this.handlePlayerTwoChange} />
-                </label>
-                <br/><br/>
-                <input type="submit" value="Submit" />
-            </form>
-        );
-    }
-}
-
-class Game extends React.Component {
-
-    constructor(props) {
-        super(props);
-        this.state = {
-            history: [{
-                squares: Array(9).fill(null),
-            }],
-            xPlayerMove: true,
-            stepNumber: 0,
-        }
-    }
-
-
-    handleClick(i) {
-        const history = this.state.history.slice(0, this.state.stepNumber + 1);
-        const current = history[history.length - 1];
-        const squares = current.squares.slice();
-        if (calculateWinner(squares) || squares[i])
-            return;
-        squares[i] = this.state.xPlayerMove ? 'X' : 'O';
-        this.setState({
-            history: history.concat([{
-                squares: squares,
-            }]),
-            xPlayerMove: !this.state.xPlayerMove,
-            stepNumber: history.length,
-        })
-    }
-
-    jumpTo(step) {
-        this.setState({
-            stepNumber: step,
-            xPlayerMove: (step % 2) === 0,
-        })
-    }
-
-    render() {
-
-        const history = this.state.history;
-        const current = history[this.state.stepNumber];
-        const winner = calculateWinner(current.squares);
-
-        const moves = history.map((step, move) => {
-            const desc = move ?
-                'Go to move #' + move :
-                'Go to game start';
-            return (
-                <li key={move}>
-                    <button onClick={() => this.jumpTo(move)}>{desc}</button>
-                </li>
-            )
-        })
-
-        let status;
-        if (winner)
-            status = "Winner:" + winner;
-        else
-            status = "Next Player: " + (this.state.xPlayerMove ? 'X' : 'O');
-
-        return (
-            <div className="game">
-                <div className="playerForm">
-                    <NameForm/>
-                </div>
-                <div className="game-board">
-                    <Board squares={current.squares}
-                        onClick={(i) => this.handleClick(i)} />
-                </div>
-                <div className="game-info">
-                    <div>{status}</div>
-                    <ol>{moves}</ol>
-                </div>
-            </div>
-        )
-    }
-}
 
 ReactDOM.render(
     <Game />,
